#933 Restructure web into testable MVC and web modules
