--- hartshorn-core/src/main/java/org/dockbox/hartshorn/proxy/Invokable.java
@@ -1,35 +1,0 @@
-/*
- * Copyright 2019-2023 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.dockbox.hartshorn.proxy;
-
-public interface Invokable {
-    Object invoke(Object obj, Object... args) throws Exception;
-
-    void setAccessible(boolean accessible);
-
-    Class<?> declaringClass();
-
-    String name();
-
-    boolean isDefault();
-
-    Class<?> returnType();
-
-    Class<?>[] parameterTypes();
-
-    String qualifiedName();
-}
