+++ hartshorn-mvc/hartshorn-mvc-freemarker/hartshorn-mvc-freemarker.gradle.kts
@@ -19,15 +19,7 @@ apply {
 }
 
 dependencies {
-    api(libs.jakartaServlet)
-
+    api(libs.freemarker)
     implementation("org.dockbox.hartshorn:hartshorn-core")
-    implementation("org.dockbox.hartshorn:hartshorn-util")
-    implementation("org.dockbox.hartshorn:hartshorn-introspect")
-    implementation("org.dockbox.hartshorn:hartshorn-events")
-    implementation("org.dockbox.hartshorn:hartshorn-config")
-
-    testImplementation(libs.httpclient)
-    testImplementation("org.dockbox.hartshorn:hartshorn-config-jackson")
-    testImplementation("org.dockbox.hartshorn:hartshorn-introspect-reflection")
+    implementation("org.dockbox.hartshorn:hartshorn-mvc")
 }
