+++ hartshorn-mvc/src/main/java/org/dockbox/hartshorn/web/mvc/template/ClassPathViewTemplate.java
@@ -14,20 +14,17 @@
  * limitations under the License.
  */
 
-apply {
-    from("${project.rootDir}/gradle/publications.gradle.kts")
-}
+package org.dockbox.hartshorn.web.mvc.template;
 
-dependencies {
-    api(libs.jakartaServlet)
+public class ClassPathViewTemplate implements ViewTemplate {
 
-    implementation("org.dockbox.hartshorn:hartshorn-core")
-    implementation("org.dockbox.hartshorn:hartshorn-util")
-    implementation("org.dockbox.hartshorn:hartshorn-introspect")
-    implementation("org.dockbox.hartshorn:hartshorn-events")
-    implementation("org.dockbox.hartshorn:hartshorn-config")
+    private final String location;
 
-    testImplementation(libs.httpclient)
-    testImplementation("org.dockbox.hartshorn:hartshorn-config-jackson")
-    testImplementation("org.dockbox.hartshorn:hartshorn-introspect-reflection")
+    public ClassPathViewTemplate(final String location) {
+        this.location = location;
+    }
+
+    public String location() {
+        return this.location;
+    }
 }
