+++ hartshorn-mvc/src/main/java/org/dockbox/hartshorn/web/mvc/template/FileViewTemplate.java
@@ -14,20 +14,19 @@
  * limitations under the License.
  */
 
-apply {
-    from("${project.rootDir}/gradle/publications.gradle.kts")
-}
+package org.dockbox.hartshorn.web.mvc.template;
 
-dependencies {
-    api(libs.jakartaServlet)
+import java.nio.file.Path;
 
-    implementation("org.dockbox.hartshorn:hartshorn-core")
-    implementation("org.dockbox.hartshorn:hartshorn-util")
-    implementation("org.dockbox.hartshorn:hartshorn-introspect")
-    implementation("org.dockbox.hartshorn:hartshorn-events")
-    implementation("org.dockbox.hartshorn:hartshorn-config")
+public class FileViewTemplate implements ViewTemplate {
 
-    testImplementation(libs.httpclient)
-    testImplementation("org.dockbox.hartshorn:hartshorn-config-jackson")
-    testImplementation("org.dockbox.hartshorn:hartshorn-introspect-reflection")
+    private final Path path;
+
+    public FileViewTemplate(final Path path) {
+        this.path = path;
+    }
+
+    public Path path() {
+        return this.path;
+    }
 }
