+++ hartshorn-proxy/src/main/java/org/dockbox/hartshorn/proxy/loaders/ObjectEqualsParameterLoaderRule.java
@@ -25,7 +25,7 @@ import org.dockbox.hartshorn.util.parameter.ParameterLoaderRule;
 public class ObjectEqualsParameterLoaderRule implements ParameterLoaderRule<ProxyParameterLoaderContext> {
     @Override
     public boolean accepts(final ParameterView<?> parameter, final int index, final ProxyParameterLoaderContext context, final Object... args) {
-        final ExecutableElementView<?> executable = parameter.declaredBy();
+        final ExecutableElementView<?, ?> executable = parameter.declaredBy();
         return executable.declaredBy().is(Object.class) && "equals".equals(executable.name());
     }
