+++ hartshorn-web/hartshorn-web-jetty/src/main/java/org/dockbox/hartshorn/web/jetty/JettyErrorHandler.java
