--- hartshorn-web/src/main/java/org/dockbox/hartshorn/web/MvcControllerContext.java
@@ -1,37 +1,0 @@
-/*
- * Copyright 2019-2023 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.dockbox.hartshorn.web;
-
-import org.dockbox.hartshorn.context.DefaultProvisionContext;
-import org.dockbox.hartshorn.context.InstallIfAbsent;
-
-import java.util.Set;
-import java.util.concurrent.ConcurrentHashMap;
-
-@InstallIfAbsent
-public class MvcControllerContext extends DefaultProvisionContext {
-
-    private final Set<RequestHandlerContext> requestHandlerContexts = ConcurrentHashMap.newKeySet();
-
-    public Set<RequestHandlerContext> requestHandlerContexts() {
-        return this.requestHandlerContexts;
-    }
-
-    public void add(final RequestHandlerContext context) {
-        this.requestHandlerContexts.add(context);
-    }
-}
