--- hartshorn-web/src/main/java/org/dockbox/hartshorn/web/mvc/MVCInitializer.java
@@ -1,32 +1,0 @@
-/*
- * Copyright 2019-2023 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.dockbox.hartshorn.web.mvc;
-
-import org.dockbox.hartshorn.application.context.ApplicationContext;
-import org.dockbox.hartshorn.util.ApplicationException;
-import org.dockbox.hartshorn.web.HttpWebServer;
-
-import java.util.Set;
-
-public interface MVCInitializer {
-
-    String TEMPLATE_ROOT = HttpWebServer.WEB_INF + "templates/";
-
-    void initialize(ApplicationContext applicationContext) throws ApplicationException;
-    String transform(ViewTemplate template, ViewModel model) throws ApplicationException;
-    Set<ViewTemplate> templates();
-}
