+++ hartshorn-web/src/main/java/org/dockbox/hartshorn/web/servlet/HandleWebServlet.java
@@ -1,0 +1,8 @@
+package org.dockbox.hartshorn.web.servlet;
+
+import org.dockbox.hartshorn.web.ServletHandler;
+
+public interface HandleWebServlet extends WebServlet {
+
+    ServletHandler handler();
+}
